+++ anyscale_provider/triggers/anyscale.py
@@ -29,7 +29,7 @@ class AnyscaleJobTrigger(BaseTrigger):
     :param poll_interval: Optional. Interval in seconds between status checks. Defaults to 60 seconds.
     """
 
-    def __init__(self, conn_id: str, job_id: str, poll_interval: int = 60):
+    def __init__(self, conn_id: str, job_id: str, poll_interval: float = 60):
         super().__init__()  # type: ignore[no-untyped-call]
         self.conn_id = conn_id
         self.job_id = job_id
@@ -113,7 +113,7 @@ class AnyscaleServiceTrigger(BaseTrigger):
         service_name: str,
         expected_state: str,
         canary_percent: float | None,
-        poll_interval: int = 60,
+        poll_interval: float = 60,
     ):
         super().__init__()  # type: ignore[no-untyped-call]
         self.conn_id = conn_id
